feat: implement synchronisation on new block arrival
Added a new property 'OnNewBlock' to identify new block broadcasts. Extended the functionality in P2PDeviceApi.cs to handle the arrival of a new block, including a throttling mechanism to manage the processing of incoming blocks. After a block is saved, the 'OnNewBlock' broadcasting event in Graph.cs is triggered.